Separate utils used everywhere into a separate file, adapt tests so they dont start breaking
